monkey_run/requirements.txt
pyxel==1.9.18
python-dotenv==1.0.0
google-generativeai==0.3.2
